+++ cmd/bnsd/scenarios/bnsdtest/bnsdtest.go
@@ -218,7 +218,7 @@ func initGenesis(t testing.TB, env *EnvConf, filename string) {
 		"escrow": []interface{}{
 			dict{
 				"sender":    "0000000000000000000000000000000000000000",
-				"arbiter":   "multisig/usage/0000000000000001",
+				"arbiter":   env.MultiSigContract.Address(),
 				"recipient": "cond:dist/revenue/0000000000000001",
 				"amount": []interface{}{
 					dict{
@@ -245,7 +245,6 @@ func initGenesis(t testing.TB, env *EnvConf, filename string) {
 			dict{"ver": 1, "pkg": "distribution"},
 			dict{"ver": 1, "pkg": "escrow"},
 			dict{"ver": 1, "pkg": "gov"},
-			dict{"ver": 1, "pkg": "hashlock"},
 			dict{"ver": 1, "pkg": "msgfee"},
 			dict{"ver": 1, "pkg": "multisig"},
 			dict{"ver": 1, "pkg": "namecoin"},