+++ x/escrow/init.go
@@ -17,17 +17,16 @@ type Initializer struct {
 // FromGenesis will parse initial escrow  info from genesis and save it in the database.
 func (i *Initializer) FromGenesis(opts weave.Options, db weave.KVStore) error {
 	var escrows []struct {
-		Sender    weave.Address   `json:"sender"`
-		Arbiter   weave.Condition `json:"arbiter"`
-		Recipient weave.Address   `json:"recipient"`
-		Timeout   weave.UnixTime  `json:"timeout"`
-		Amount    []*coin.Coin    `json:"amount"`
+		Sender    weave.Address  `json:"sender"`
+		Arbiter   weave.Address  `json:"arbiter"`
+		Recipient weave.Address  `json:"recipient"`
+		Timeout   weave.UnixTime `json:"timeout"`
+		Amount    []*coin.Coin   `json:"amount"`
 	}
 
 	if err := opts.ReadOptions("escrow", &escrows); err != nil {
 		return err
 	}
-
 	bucket := NewBucket()
 	for j, e := range escrows {
 		escr := Escrow{
@@ -37,6 +36,7 @@ func (i *Initializer) FromGenesis(opts weave.Options, db weave.KVStore) error {
 			Recipient: e.Recipient,
 			Timeout:   e.Timeout,
 		}
+
 		if err := escr.Validate(); err != nil {
 			return errors.Wrapf(err, "invalid escrow at position: %d ", j)
 		}